Fix build for distros with more than one ruby interpreter
In most of the openSUSE Distros, we carry only one ruby interpreter.
openSUSE Tumbleweed is special though; as a rolling distro it often
tries to accomodate users to changes. It's planned to add ruby 2.7 soon
(the interpreter is already present) and enable rubygems to build
for ruby 2.6 and 2.7 and later disable 2.6.

For yast, it does not make sense to provided gems for both versions and
we just build for the distro defined default version.